@@ -1,34 +1,27 @@
 package com.nirsb.oppa.oppaservice.model;
-
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
+
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
-@ApiModel(description = "It contains Document Information")
 @Component
-public class Document{
-    @ApiModelProperty(notes = "Users' mobile number to notify him")
-    @Size(min = 10, max = 10,message = "Mobile number size should be 10")
+public class Document {
+
+    @Size(min = 10, max = 10, message = "Mobile number size should be 10")
     private long mobileNumber;
-    @ApiModelProperty(notes = "How many copy user wants to print")
+
     @NotBlank(message = "Number of copy can not be blank")
     private int noOfCopies;
-    @ApiModelProperty(notes = "Color or black & White")
     @NotNull
     private String docTypes;
-    @ApiModelProperty(notes = "OneSided or DoubleSided")
+
     @NotNull
     private String docFormat;
-    @ApiModelProperty(notes = "Users' email address to notify him about print status")
+
     @Email(message = "Please enter a valid email address")
     private String emailAddress;
-    @ApiModelProperty(notes = "Document which user want to print")
-    private MultipartFile file;
 
     public Document() {
     }
@@ -74,13 +67,6 @@ public class Document{
         this.docFormat = docFormat;
     }
 
-    public MultipartFile getFile() {
-        return file;
-    }
-
-    public void setFile(MultipartFile file) {
-        this.file = file;
-    }
 
     @Override
     public String toString() {
