@@ -1,6 +1,6 @@
 package com.nirsb.oppa.oppaservice;
 
-import org.junit.Test;
+import org.testng.annotations.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
